halfway there

+++ src/main/java/com/sammy/lodestone/helpers/BlockHelper.java
@@ -428,7 +428,7 @@ public class BlockHelper {
 		ArrayList<BlockPos> positions = new ArrayList<>();
 		for (int i = 0; i < nodes; i++) {
 			PathNode node = path.getNode(i);
-			positions.add(new BlockPos(node.x, node.y - 0.5, node.z));
+			positions.add(new BlockPos(node.x, (int) (node.y - 0.5), node.z));
 		}
 		if (!inclusive) {
 			positions.remove(0);
